% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_material_side_nav_tab.R
\name{activate_material_side_nav_tab}
\alias{activate_material_side_nav_tab}
\title{Update the active side nav tab from the code.}
\usage{
activate_material_side_nav_tab(side_nav_tab_id)
}
\arguments{
\item{side_nav_tab_id}{The side_nav_tab_id of the material_side_nav_tab_content to be shown.}
}
\description{
Update the active side nav tab from the code.
}
\examples{
\dontrun{
activate_material_side_nav_tab(
  input_id = "example_material_side_nav_tab_content",
)
}
}
\seealso{
\code{\link{material_side_nav_tab_content}}
}
