% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close-shiny-material-modal.R
\name{close_material_modal}
\alias{close_material_modal}
\title{Close a material modal programmatically.}
\usage{
close_material_modal(session, modal_id)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{modal_id}{String. The ID of the modal to open.}
}
\description{
Close a material modal programmatically (server side).
}
\examples{
\dontrun{
close_material_modal(session, "example_modal")
}
}
