% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-card.R
\name{material_card}
\alias{material_card}
\title{Create a card that will contain UI content}
\usage{
material_card(title, ...)
}
\arguments{
\item{title}{String. The title of the card}

\item{...}{The UI elements to place in the card}
}
\description{
UI content can be placed in cards to organize items on a page.
}
\examples{
material_card(
  title = "Example Card",
  shiny::tags$h1("Card Content")
)
}
