% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-checkbox.R
\name{material_checkbox}
\alias{material_checkbox}
\title{Create a shinymaterial checkbox}
\usage{
material_checkbox(input_id, label, initial_value = FALSE, color = NULL)
}
\arguments{
\item{input_id}{String. The input identifier used to access the value.}

\item{label}{String. The checkbox label.}

\item{initial_value}{Boolean. Is the checkbox initially checked?}

\item{color}{String. The color of the check. Leave empty for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors. \emph{This input requires using color hex codes, rather than the word form. E.g., "#ef5350", rather than "red lighten-1".}}
}
\description{
Build a shinymaterial checkbox. The value is a boolean (TRUE if checked, FALSE if not checked).
}
