% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-collapsible.R
\name{material_collapsible}
\alias{material_collapsible}
\title{Create a collapsible that will contain collapsible items}
\usage{
material_collapsible(..., depth = NULL, type = NULL)
}
\arguments{
\item{...}{The collapsible item elements}

\item{depth}{Integer. The amount of depth of the card. The value should be between 0 and 5. Leave empty for the default depth.}

\item{type}{String. By Default : "accordion". Can also use c("expandable", "popout").}
}
\description{
Collapsible items must be placed in collapsible.
}
\examples{
material_collapsible(
  type = "popout",
  material_collapsible_item(label = "First", icon = "filter_drama"),
  material_collapsible_item(label = "Second", icon = "place"),
  material_collapsible_item(label = "Third", icon = "whatshot")
)
}
\seealso{
\code{\link{material_collapsible_item}}
}
