% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-collapsible-item.R
\name{material_collapsible_item}
\alias{material_collapsible_item}
\title{Create a collapsible item that will contain UI content}
\usage{
material_collapsible_item(label, ..., icon = NULL, active = FALSE)
}
\arguments{
\item{label}{String. The collapsible item label.}

\item{...}{The UI elements to place in the collapsible item.}

\item{icon}{String. The name of the icon. Leave empty for no icon. Visit \url{http://materializecss.com/icons.html} for a list of available icons.}

\item{active}{Boolean. By default FALSE. if TRUE, the collapsible item will be active.}
}
\description{
UI content can be placed in collapsible item to organize items on a page.
}
\examples{
material_collapsible_item(
 label = "First", 
 icon = "filter_drama",
 active = TRUE,
 material_button("test", label = "test")
)
}
\seealso{
\code{\link{material_collapsible}}
}
