% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-color.R
\docType{data}
\name{material_color}
\alias{material_color}
\title{Available material colors}
\format{An object of class \code{list} of length 254.}
\usage{
material_color
}
\description{
A list of colors from the material pallete.
}
\examples{
material_checkbox(
  input_id = "example_checkbox",
  label = "Checkbox",
  initial_value = TRUE,
  color = material_color$red_accent_1
)
}
\seealso{
\url{http://materializecss.com/color.html}
}
\keyword{datasets}
