% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-date-picker.R
\name{material_date_picker}
\alias{material_date_picker}
\title{Create a shinymaterial date picker}
\usage{
material_date_picker(input_id, label, value = "", color = NULL)
}
\arguments{
\item{input_id}{String. The input identifier used to access the value.}

\item{label}{String. The date picker label.}

\item{value}{String. The starting date (format 'mmm dd, yyyy').}

\item{color}{String. The date picker color. Leave empty for the default color. Visit \url{https://materializecss.com/color.html} for a list of available colors.}
}
\description{
Build a shinymaterial date picker.
}
\examples{
material_date_picker(
  input_id = "example_date_picker",
  label = "Date Picker",
  value = "Jun 02, 2006"
)
}
\seealso{
\code{\link{update_material_date_picker}}
}
