% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_depth.R
\name{material_depth}
\alias{material_depth}
\title{Add depth to a UI element}
\usage{
material_depth(..., depth = 4)
}
\arguments{
\item{...}{The UI elements to place in the column}

\item{depth}{Integer. The amount of depth. The value should be between 0 and 5. A value of 0 can be used to remove depth from objects that have depth by default.}
}
\description{
Give a UI element an illusion of depth by creating a shadow.
}
\examples{
material_depth(
  depth = 5,
  material_card(title = "Example Depth")
)
}
