% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-input.R
\name{material_input}
\alias{material_input}
\title{Create a shinymaterial input}
\usage{
material_input(type, ...)
}
\arguments{
\item{type}{String. The type of input to be created.}

\item{...}{Additional arguments for the input.}
}
\description{
Build a shinymaterial input.
}
\examples{
##-- switch --##
material_input(
  type = "switch",
  input_id = "example_switch",
  off_label = "Off",
  on_label = "On",
  initial_value = TRUE
)
}
