% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-modal.R
\name{material_modal}
\alias{material_modal}
\title{Place UI content in a modal}
\usage{
material_modal(modal_id, modal_button_text, modal_title, ...)
}
\arguments{
\item{modal_id}{String. The ID for the modal. Must be unique per app}

\item{modal_button_text}{String. The text displayed on the modal trigger button}

\item{modal_title}{String. The title of the modal window}

\item{...}{The UI elements to place in the modal}
}
\description{
Put any UI object inside of a modal. The modal will open when the button is pressed.
}
\examples{
material_modal(
  modal_id = "example_modal",
  modal_button_text = "Modal",
  modal_title = "Example Modal Title",
  shiny::tags$p("Modal Content")
)
}
