% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-page.R
\name{material_page}
\alias{material_page}
\title{Create a shinymaterial page}
\usage{
material_page(title, ...)
}
\arguments{
\item{title}{String. The title of the page.}

\item{...}{The UI elements to place in the page}
}
\description{
Build a shinymaterial page.
}
\examples{
material_page(
  title = "Example Title",
  shiny::tags$h1("Page Content")
)
}
