% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-parallax.R
\name{material_parallax}
\alias{material_parallax}
\title{Create a parallax image}
\usage{
material_parallax(image_source)
}
\arguments{
\item{image_source}{String. The image file name. Place the image in a folder labeled 'www' at the same level as the application (server.R & ui.R)}
}
\description{
Use this function to create a parallax effect in your application.
}
\examples{
material_parallax(
  image_source = "example_image.jpg"
)
}
