% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-password-box.R
\name{material_password_box}
\alias{material_password_box}
\title{Create a shinymaterial password box}
\usage{
material_password_box(input_id, label, color = NULL)
}
\arguments{
\item{input_id}{String. The input identifier used to access the value.}

\item{label}{String. The password box label.}

\item{color}{String. The accent color of the password box. Leave empty for the default color.}
}
\description{
Build a shinymaterial password box.
}
