% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-row.R
\name{material_row}
\alias{material_row}
\title{Create a row to organize UI content}
\usage{
material_row(...)
}
\arguments{
\item{...}{The UI elements to place in the row.}
}
\description{
UI content can be placed in a row to organize items on a page.
}
\examples{
material_row(
  shiny::tags$h1("Row Content")
)
}
