% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-side-nav.R
\name{material_side_nav}
\alias{material_side_nav}
\title{Create a side-nav that contains UI content}
\usage{
material_side_nav(..., fixed = FALSE)
}
\arguments{
\item{...}{The UI elements to place in the side-nav}

\item{fixed}{A boolean. Set to TRUE to keep side-nav open on large screens.}
}
\description{
UI content can be placed in side-nav.
}
\examples{
material_side_nav(
  fixed = FALSE,
  shiny::tags$h1("Side-nav Content")
)
}
