% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-side-nav-tab-content.R
\name{material_side_nav_tab_content}
\alias{material_side_nav_tab_content}
\title{Place UI content within a side-nav tab}
\usage{
material_side_nav_tab_content(side_nav_tab_id, ...)
}
\arguments{
\item{...}{The UI elements to place in the side nav tab.}

\item{tab_id}{String. The side-nav tab id in which to place the UI content.}
}
\description{
Use this function to place UI content within a specific side-nav tab.
}
\examples{
material_side_nav_tab_content(
  tab_id = "example_tab_1",
  shiny::tags$h1("Side-Nav Tab Content")
)
}
