% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-side-nav-tabs.R
\name{material_side_nav_tabs}
\alias{material_side_nav_tabs}
\title{Place UI content within a side-nav tab}
\usage{
material_side_nav_tabs(side_nav_tabs, color = NULL)
}
\arguments{
\item{color}{String. The accent color of the tabs. Leave blank for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors. \emph{Tab color requires using word forms of colors (e.g. 'deep-purple'). Also, lighten or darken effects do not work on tab colors.}}

\item{tabs}{Named vector. The tab display names and corresponding side-nav tab ids.}
}
\description{
Use this function to create side-nav tabs in your application.
}
\examples{
material_side_nav_tabs(
  side_nav_tabs = c(
    "Example Side-Nav Tab 1" = "example_side_nav_tab_1",
    "Example Side-Nav Tab 2" = "example_side_nav_tab_2"
  ),
  color = "deep-purple"
)
}
