% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-tabs.R
\name{material_tabs}
\alias{material_tabs}
\title{Place UI content within a tab}
\usage{
material_tabs(tabs, in_card = FALSE)
}
\arguments{
\item{tabs}{Named Vector. The tab display names as well as the tab ids.}

\item{in_card}{Boolean. Are the tabs within a card?}
}
\description{
Use this function to create tabs in your application.
}
\examples{
material_tabs(
  tabs = c(
    "Example Tab 1" = "example_tab_1",
    "Example Tab 2" = "example_tab_2"
  )
)
}
