% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-text-box.R
\name{material_text_box}
\alias{material_text_box}
\title{Create a shinymaterial text box}
\usage{
material_text_box(input_id, label, color = NULL)
}
\arguments{
\item{input_id}{String. The input identifier used to access the value.}

\item{label}{String. The text box label.}

\item{color}{String. The accent color of the text box. Leave empty for the default color.}
}
\description{
Build a shinymaterial text box.
}
