% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-ui-output.R
\name{material_ui_output}
\alias{material_ui_output}
\title{Display a shinymaterial output element}
\usage{
material_ui_output(material_output_id)
}
\arguments{
\item{material_output_id}{Output id from rendered material UI.}
}
\description{
This function is used on the UI side to display an element rendered using render_material_ui().
}
\details{
material_ui_output() is intended to be used with render_material_ui() on the server side.
}
\examples{
material_ui_output("renderedDropdown")
}
\seealso{
\code{\link{render_material_ui}}
}
