% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-shiny-material-modal.R
\name{open_material_modal}
\alias{open_material_modal}
\title{Open a material modal programmatically.}
\usage{
open_material_modal(session, modal_id)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{modal_id}{String. The ID of the modal to open.}
}
\description{
Open a material modal programmatically (server side).
}
\examples{
open_material_modal(session, "example_modal")
}
