% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-shiny-material-from-server.R
\name{render_material_from_server}
\alias{render_material_from_server}
\title{Render reactive UI shinymaterial elements}
\usage{
render_material_from_server(material_ui)
}
\arguments{
\item{material_ui}{shinymaterial UI element(s)}
}
\description{
This function is used within a shiny::renderUI(). The corresponding output is referenced using shiny::uiOutput().
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- material_page(
  uiOutput("renderedButton")
)

server <- function(input, output) {
  output$renderedButton <- renderUI({
  render_material_from_server(material_button("example_button", "Button"))
})
}
shinyApp(ui, server)
}
}
