% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-shiny-material-ui.R
\name{render_material_ui}
\alias{render_material_ui}
\title{Renders reactive UI elements using the shinymaterial package}
\usage{
render_material_ui(material_ui)
}
\arguments{
\item{material_ui}{An expression that returns shinymaterial UI elements.}
}
\description{
This function is used on the server side to render an element that is to be displayed using material_ui_output().
}
\details{
render_material_ui() is intended to be used with material_ui_output() on the UI side.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {

ui <- material_page(
  material_ui_output("renderedDropdown")
)

server <- function(input, output) {
  output$renderedDropdown <- render_material_ui({
  material_dropdown(
                    input_id = "renderedDropdown",
                    label = "Selection",
                    choices = c(1,2)
                    )
})
}
shinyApp(ui, server)
}
}
\seealso{
\code{\link{material_ui_output}}
}
