% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-side-nav-tabs-info.R
\name{side_nav_tabs_info}
\alias{side_nav_tabs_info}
\title{Query information about the side-nav tabs.}
\usage{
side_nav_tabs_info(input = NULL)
}
\arguments{
\item{input}{The input object in the shiny session.}
}
\description{
Query information about the side-nav tabs.
}
\examples{
side_nav_tabs_info(input)
}
