% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-shiny-material-checkbox.R
\name{update_material_checkbox}
\alias{update_material_checkbox}
\title{Change the value of a material_checkbox on the client}
\usage{
update_material_checkbox(session, input_id, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{input_id}{The input_id of the material_checkbox.}

\item{value}{Boolean. The value to set for the material_checkbox.}
}
\description{
Change the value of a material_checkbox on the client.
}
\examples{
\dontrun{
update_material_checkbox(
  session,
  input_id = "example_checkbox",
  value = TRUE
)
}
}
