% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-shiny-material-date-picker.R
\name{update_material_date_picker}
\alias{update_material_date_picker}
\title{Change the value of a material_date_picker on the client}
\usage{
update_material_date_picker(session, input_id, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{input_id}{The input_id of the material_date_picker.}

\item{value}{The value to set for the material_date_picker.}
}
\description{
Change the value of a material_date_picker on the client.
}
\examples{
\dontrun{
update_material_date_picker(
  session,
  input_id = "example_date_picker",
  value = "10 April, 2012"
)
}
}
\seealso{
\code{\link{material_date_picker}}
}
