% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-shiny-material-dropdown.R
\name{update_material_dropdown}
\alias{update_material_dropdown}
\title{Change the value of a material_dropdown on the client}
\usage{
update_material_dropdown(session, input_id, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{input_id}{The input_id of the material_dropdown.}

\item{value}{The value to set for the material_dropdown.}
}
\description{
Change the value of a material_dropdown on the client.
}
\examples{
\dontrun{
update_material_dropdown(
  session,
  input_id = "example_dropdown",
  value = "New Text"
)
}
}
