% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-shiny-material-radio-button.R
\name{update_material_radio_button}
\alias{update_material_radio_button}
\title{Change the value of a material_radio_button on the client}
\usage{
update_material_radio_button(session, input_id, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{input_id}{The input_id of the material_radio_button.}

\item{value}{The value to set for the material_radio_button.}
}
\description{
Change the value of a material_radio_button on the client.
}
\examples{
\dontrun{
update_material_radio_button(
  session,
  input_id = "example_radio_button",
  value = "new_value"
)
}
}
\seealso{
\code{\link{material_radio_button}}
}
