% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-shiny-material-slider.R
\name{update_material_slider}
\alias{update_material_slider}
\title{Change the value of a material_slider on the client}
\usage{
update_material_slider(session, input_id, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{input_id}{The input_id of the material_slider.}

\item{value}{The value to set for the material_slider.}
}
\description{
Change the value of a material_slider on the client.
}
\examples{
\dontrun{
update_material_slider(
  session,
  input_id = "example_slider",
  value = "new_value"
)
}
}
\seealso{
\code{\link{material_slider}}
}
