% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-shiny-material-switch.R
\name{update_material_switch}
\alias{update_material_switch}
\title{Change the value of a material_switch on the client}
\usage{
update_material_switch(session, input_id, value = NULL)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{input_id}{The input_id of the material_switch.}

\item{value}{Boolean. The value to set for the material_switch.}
}
\description{
Change the value of a material_switch on the client.
}
\examples{
\dontrun{
update_material_switch(
  session,
  input_id = "example_switch",
  value = TRUE
)
}
}
