% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-file-input.R
\name{material_file_input}
\alias{material_file_input}
\title{Create a shinymaterial file input}
\usage{
material_file_input(input_id, label = "File", color = NULL)
}
\arguments{
\item{input_id}{String. The input identifier used to access the value.}

\item{label}{String. The button text.}

\item{color}{String. The color of the button. Leave empty for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{icon}{String. The name of the icon. Leave empty for no icon. Visit \url{http://materializecss.com/icons.html} for a list of available icons.}

\item{depth}{Integer. The amount of depth of the file input button. The value should be between 0 and 5. Leave empty for the default depth.}
}
\description{
Build a shinymaterial file input.
}
\examples{
material_file_input(
  input_id = "example_file_input",
  label = "File"
)
}
