% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-number-box.R
\name{material_number_box}
\alias{material_number_box}
\title{Create a shinymaterial number box}
\usage{
material_number_box(
  input_id,
  label,
  min_value,
  max_value,
  step_size = 1,
  initial_value,
  color = NULL
)
}
\arguments{
\item{input_id}{String. The input identifier used to access the value.}

\item{label}{String. The number box label.}

\item{min_value}{Number. The minimum allowable value.}

\item{max_value}{Number. The maximum allowable value.}

\item{step_size}{Number. The step size of the arrow clicks.}

\item{initial_value}{Number. The initial value.}

\item{color}{String. The accent color of the number box. Leave empty for the default color. Visit \url{http://materializecss.com/color.html} for a list of available colors. \emph{This input requires using color hex codes, rather than the word form. E.g., "#ef5350", rather than "red lighten-1".}}
}
\description{
Build a shinymaterial number box.
}
\examples{
material_number_box(
  input_id = "example_number_box",
  label = "number box",
  min_value = 5,
  max_value = 15,
  initial_value = 10,
  step_size = 2,
  color = "#ef5350"
)
}
\seealso{
\code{\link{update_material_number_box}}
}
