% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-page.R
\name{material_page}
\alias{material_page}
\title{Create a shinymaterial page}
\usage{
material_page(..., title = "", nav_bar_color = NULL,
  background_color = "grey lighten-4", font_color = NULL)
}
\arguments{
\item{...}{The UI elements to place in the page.}

\item{title}{String. The title of the page.}

\item{nav_bar_color}{Color of the navigation bar. Leave blank for the default color.}

\item{background_color}{Page background color. Leave blank for the default color.}

\item{font_color}{String. The title font color. Leave blank for the default color. \emph{Title color requires using word forms of colors (e.g. "deep-purple"). Also, lighten or darken effects do not work on title colors.}}
}
\description{
Build a shinymaterial page.
}
\examples{
material_page(
  title = "Example Title",
  nav_bar_color = "red lighten-2",
  background_color = "blue lighten-4",
  shiny::tags$h1("Page Content")
)
}
\seealso{
\code{\link{material_color}} for a list of available colors.
}
