% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-radio-button.R
\name{material_radio_button}
\alias{material_radio_button}
\title{Create a shinymaterial radio button}
\usage{
material_radio_button(input_id, label, choices, color = NULL)
}
\arguments{
\item{input_id}{String. The input identifier used to access the value.}

\item{label}{String. The radio button label.}

\item{choices}{Named vector. The option names and underyling values.}

\item{color}{String. The color of the radio buttons. Leave empty for the default color.}
}
\description{
Build a shinymaterial radio button.
}
\examples{
material_radio_button(
  input_id = "example_radio_button",
  label = "Radio Button",
  choices = c(
    "Cake" = "c",
    "Pie" = "p",
    "Brownie" = "b"
  ),
  color = "blue"
)
}
