% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-tabs.R
\name{material_tabs}
\alias{material_tabs}
\title{Place UI content within a tab}
\usage{
material_tabs(tabs, color = NULL)
}
\arguments{
\item{tabs}{Named vector. The tab display names and corresponding tab ids.}

\item{color}{String. The accent color of the tabs. Leave blank for the default color. Must be valid css color.}
}
\description{
Use this function to create tabs in your application.
}
