% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-shiny-material-button.R
\name{update_material_button}
\alias{update_material_button}
\title{Change the text, icon of a material_button on the client. Allow to disable.}
\usage{
update_material_button(session, input_id, label = NULL, icon = NULL,
  enable = TRUE)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{input_id}{The input_id of the material_button.}

\item{label}{The new label of the material_button.}

\item{icon}{The new icon of the material_button. If not set, icon disappear.}

\item{enable}{TRUE if the button is enable, FALSE if disable.}
}
\description{
Change the value text, icon of a material_button on the client. 
Allow to disable the button and then enable.
}
\examples{
\dontrun{
update_material_text_box(
  session,
  input_id = "example_button",
  value = "New Text",
  icon = "stop",
  enable = FALSE
)
}
}
